[Fix] PR Review 반영하여 User Defalut extension파일을 삭제하고 실제 파일에 기능추가
